mailbox: ti-msgmgr: Use IRQF_NO_SUSPEND flag for wakeup interrupts
The ti-msgmgr interrupt is shared and triggers also for the padconf
wake-up events from deeper idle states. We need to allow interrupts to
trigger always with IRQF_NO_SUSPEND. This flag is also set for
pinctrl-single that handles the padconf wake-up events.

This should be also OK to enable for ti,k2g-message-manager in addition
to ti,am654-secure-proxy even though the features are not used. If
needed, we can set IRQF_NO_SUSPEND based on the compatible property.

Note that eventually we may want to set up ti-msgmgr as a chained
interrupt handler that pinctrl-single can use. This allows filtering out
the unnecessary mailbox interrupts.

Signed-off-by: Tony Lindgren <tony@atomide.com>