HACK: arm64: dts: ti: Configure am62 pinctrl-single for wake-up interrupts
Configure am62 pinctrl-single for wake-up interrupts.

Eventually we may want to set up a chained interrupt handler in the
ti-msgmgr for pinctrl-single to use.

Marked as HACK because physically none of the wakeup source events are
connected to GIC.

[d-gole@ti.com] change to use the GIC_SPI 98 instead of 34. The logic is
such that in DT we specify x+32 for configuring intr source gicss0.spi.x
In this case the input index is gicss0.spi.66 so interrupt source will be
dmss0.intaggr_vintr.2

Co-developed-by: Dhruva Gole <d-gole@ti.com>
Signed-off-by: Dhruva Gole <d-gole@ti.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>