WIP: AM62L: power-domain-map use limited set of sudomains
Restrict the power-domain-map to a small subset of subdomains.

Since the MAIN domain transitions will only happen when all subdomains
and their devices are idle, this makes easier to trigger domain state
transistion for the MAIN domain because only the subdomains listed have
to transition to off before MAIN can transition to off.