FROMLIST: pmdomain: Respect the system-wakeup QoS limit at system-wide suspend
A system-wakeup QoS limit may have been requested by user-space. To avoid
entering a too deep state for PM domains that are managed my genpd, let's
start to take into account the QoS limit when selecting the
low-power-state.

If it turns out that none of the states in the list of domain-idlestates
are suitable to use, let's leave the PM domain powered-on.

Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Link: https://lore.kernel.org/r/20250716123323.65441-3-ulf.hansson@linaro.org
Signed-off-by: Kevin Hilman <khilman@baylibre.com>