TI: arm64: dts: ti: update am62l3-lpmdemo to enable more peripherals
Inheriting from the actual k3-am62l3-evm.dts means we now have far
lesser a delta between the two. We are now keeping the lpmdemo mainly
to demonstrate the s2idle LPM selection feature which is still a work
in progress for 2 CPUs. Hence enable only 1 CPU for now in this DTS.
Add the idle-states corresponding to the various low power modes
supported like deepsleep and RTC+DDR.

Signed-off-by: Dhruva Gole <d-gole@ti.com>